+++ R/game_systems.R
@@ -23,7 +23,11 @@
 #' \item{dice_d8}{Eight-sided dice in six color schemes (color controlled by their suit).
 #'                Octahedrons with the rank as a numeral at the top face.}
 #' \item{dice_d10}{Ten-sided dice in six color schemes (color controlled by their suit).
-#'                Pentagonal trapezohedrons with the rank as a numeral at the top face.}
+#'                Pentagonal trapezohedrons with the rank as a numeral at the top face.
+#'                The rank of ten is represented by a zero.}
+#' \item{dice_d10_percentile}{Ten-sided dice in six color schemes (color controlled by their suit).
+#'                Pentagonal trapezohedrons with the rank as a numeral followed by a zero at the top face.
+#'                The rank of ten is represented by a zero.}
 #' \item{dice_d12}{Twelve-sided dice in six color schemes (color controlled by their suit).
 #'                Dodecahedrons with the rank as a numeral at the top face.}
 #' \item{dice_d20}{Twenty-sided dice in six color schemes (color controlled by their suit).
@@ -159,6 +163,7 @@ game_systems <- function(style = NULL, round = FALSE, pawn = "token") {
          dice_d4 = dice_d4(style, color_list),
          dice_d8 = dice_d8(style, color_list),
          dice_d10 = dice_d10(style, color_list),
+         dice_d10_percentile = dice_d10_percentile(style, color_list),
          dice_d12 = dice_d12(style, color_list),
          dice_d20 = dice_d20(style, color_list),
          dice_fudge = dice_fudge(color_list, rect_shape),
@@ -340,6 +345,33 @@ dice_d10 <- function(style = "sans", color_list = color_list_fn()) {
                       background_color = "white,white,white,white,black,black",
                       shape.die = "kite",
                       shape_r.die = 0.5 - 0.1909830056250526320039,
+                      invert_colors = TRUE)
+    dice <- pp_cfg(c(dice_list, color_list))
+    dice$has_piecepack <- FALSE
+    dice$has_dice <- TRUE
+    dice
+}
+
+dice_d10_percentile <- function(style = "sans", color_list = color_list_fn()) {
+    dice_list <- list(n_suits = 6, n_ranks = 10,
+                      rank_text.die = "1,2,3,4,5,6,7,8,9,0",
+                      dm_text.die = "0",
+                      dm_cex.die = 1.15,
+                      dm_r.die = -0.05,
+                      dm_t.die = 180,
+                      ps_cex.die = 1.00,
+                      ps_r.die = 0.20,
+                      ps_t.die = 180,
+                      fontfamily = ifelse(grepl("^dejavu", style), "DejaVu Sans", "sans"),
+                      op_grob_fn.die = d10TopGrob,
+                      obj_fn.die = save_d10_obj,
+                      width.die =  5 / 8,
+                      height.die =  0.4913446110983896164548,
+                      depth.die =  0.5621585749837085810299,
+                      background_color = "white,white,white,white,black,black",
+                      shape.die = "kite",
+                      shape_r.die = 0.5 - 0.1909830056250526320039,
+                      shape_t.die = 0,
                       invert_colors = TRUE)
     dice <- pp_cfg(c(dice_list, color_list))
     dice$has_piecepack <- FALSE